# frontend/api.yaml
openapi: 3.0.0
info:
  title: GIBDD API
  version: 1.0.0
paths:
...
  /vehicles:
    get:
      summary: Get a list of vehicles
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: owner_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: A list of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vehicle'
    post:
      summary: Create a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '201':
          description: The created vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
  /vehicles/{id}:
    get:
      summary: Get a vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
    put:
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleInput'
      responses:
        '200':
          description: The updated vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
    delete:
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
...
  /thefts:
    get:
      summary: Get a list of thefts
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: A list of thefts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Theft'
    post:
      summary: Create a new theft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TheftInput'
      responses:
        '201':
          description: The created theft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theft'
  /thefts/{id}:
    get:
      summary: Get a theft by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The theft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theft'
    put:
      summary: Update a theft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TheftInput'
      responses:
        '200':
          description: The updated theft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Theft'
    delete:
      summary: Delete a theft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
  /wanted:
    get:
      summary: Get a list of wanted vehicles
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: A list of wanted vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Wanted'
    post:
      summary: Create a new wanted vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WantedInput'
      responses:
        '201':
          description: The created wanted vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wanted'
  /wanted/{id}:
    get:
      summary: Get a wanted vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The wanted vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wanted'
    put:
      summary: Update a wanted vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WantedInput'
      responses:
        '200':
          description: The updated wanted vehicle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wanted'
    delete:
      summary: Delete a wanted vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No content
...
components:
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: string
        licensePlate:
          type: string
        ownerId:
          type: string
        organizationId:
          type: string
    VehicleInput:
      type: object
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: string
        licensePlate:
          type: string
        ownerId:
          type: string
        organizationId:
          type: string
...
    Theft:
      type: object
      properties:
        id:
          type: string
        vehicleId:
          type: string
        theftDate:
          type: string
        location:
          type: string
        description:
          type: string
    TheftInput:
      type: object
      properties:
        vehicleId:
          type: string
        theftDate:
          type: string
        location:
          type: string
        description:
          type: string
    Wanted:
      type: object
      properties:
        id:
          type: string
        vehicleId:
          type: string
        additionDate:
          type: string
        reason:
          type: string
        status:
          type: string
    WantedInput:
      type: object
      properties:
        vehicleId:
          type: string
        additionDate:
          type: string
        reason:
          type: string
        status:
          type: string
...